重新生成代码，并修改相关代码

+++ onl-api/src/main/java/rebue/onl/mo/OnlOnlineSpecLogMo.java
@@ -121,17 +121,17 @@ public class OnlOnlineSpecLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    private Integer currentOnlineCount;
+    private BigDecimal currentOnlineCount;
 
     /**
-     *    限制购买数量(默认为0，不限制)
-     *                每个人限制购买的数量
+     *     限制购买数量(默认为0，不限制)
+     * 每个人限制购买的数量
      *
-     *    数据库字段: ONL_ONLINE_SPEC_LOG.LIMIT_COUNT
+     *     数据库字段: ONL_ONLINE_SPEC_LOG.LIMIT_COUNT
      *
-     *    @mbg.generated 自动生成，如需修改，请删除本行
+     *     @mbg.generated 自动生成，如需修改，请删除本行
      */
-    private Integer limitCount;
+    private BigDecimal limitCount;
 
     /**
      *    排序号
@@ -405,7 +405,7 @@ public class OnlOnlineSpecLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public Integer getCurrentOnlineCount() {
+    public BigDecimal getCurrentOnlineCount() {
         return currentOnlineCount;
     }
 
@@ -416,31 +416,31 @@ public class OnlOnlineSpecLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public void setCurrentOnlineCount(Integer currentOnlineCount) {
+    public void setCurrentOnlineCount(BigDecimal currentOnlineCount) {
         this.currentOnlineCount = currentOnlineCount;
     }
 
     /**
-     *    限制购买数量(默认为0，不限制)
-     *                每个人限制购买的数量
+     *     限制购买数量(默认为0，不限制)
+     * 每个人限制购买的数量
      *
-     *    数据库字段: ONL_ONLINE_SPEC_LOG.LIMIT_COUNT
+     *     数据库字段: ONL_ONLINE_SPEC_LOG.LIMIT_COUNT
      *
-     *    @mbg.generated 自动生成，如需修改，请删除本行
+     *     @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public Integer getLimitCount() {
+    public BigDecimal getLimitCount() {
         return limitCount;
     }
 
     /**
-     *    限制购买数量(默认为0，不限制)
-     *                每个人限制购买的数量
+     *     限制购买数量(默认为0，不限制)
+     * 每个人限制购买的数量
      *
-     *    数据库字段: ONL_ONLINE_SPEC_LOG.LIMIT_COUNT
+     *     数据库字段: ONL_ONLINE_SPEC_LOG.LIMIT_COUNT
      *
-     *    @mbg.generated 自动生成，如需修改，请删除本行
+     *     @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public void setLimitCount(Integer limitCount) {
+    public void setLimitCount(BigDecimal limitCount) {
         this.limitCount = limitCount;
     }
