+++ onl-svr-feign/src/main/java/rebue/onl/svr/feign/OnlOnlineSpecSvc.java
@@ -1,6 +1,8 @@
 package rebue.onl.svr.feign;
 
+import java.math.BigDecimal;
 import java.util.List;
+
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -21,33 +23,33 @@ import rebue.sbs.feign.FeignConfig;
 @FeignClient(name = "onl-svr", configuration = FeignConfig.class)
 public interface OnlOnlineSpecSvc {
 
-	/**
-	 * 获取单个上线规格
-	 */
-	@GetMapping("/onl/onlinespec/getbyid")
-	OnlOnlineSpecMo getById(@RequestParam("id") Long id);
+    /**
+     * 获取单个上线规格
+     */
+    @GetMapping("/onl/onlinespec/getbyid")
+    OnlOnlineSpecMo getById(@RequestParam("id") Long id);
 
-	/**
-	 * 根据上线规格id修改销售数量(减)
-	 */
-	@PutMapping(value = "/onl/onlinespec/modifysalecountbyid")
-	Ro modifySaleCountById(@RequestParam("id") Long id, @RequestParam("buyCount") Integer buyCount);
+    /**
+     * 根据上线规格id修改销售数量(减)
+     */
+    @PutMapping(value = "/onl/onlinespec/modifysalecountbyid")
+    Ro modifySaleCountById(@RequestParam("id") Long id, @RequestParam("buyCount") BigDecimal buyCount);
 
-	/**
-	 * 获取上线规格信息 Title: selectOnlineSpecInfoByOnlineId Description:
-	 */
-	@GetMapping(value = "/onl/onlinespec/details")
-	List<OnlOnlineSpecInfoRo> selectOnlineSpecInfo(@RequestParam("onlineId") Long onlineId,
-			@RequestParam("onlineSpec") String onlineSpec);
+    /**
+     * 获取上线规格信息 Title: selectOnlineSpecInfoByOnlineId Description:
+     */
+    @GetMapping(value = "/onl/onlinespec/details")
+    List<OnlOnlineSpecInfoRo> selectOnlineSpecInfo(@RequestParam("onlineId") Long onlineId,
+            @RequestParam("onlineSpec") String onlineSpec);
 
-	/**
-	 * 根据上线规格id修改是否已有首单
-	 * 
-	 * @param id
-	 * @param isHaveFirstOrder
-	 * @return
-	 */
-	@PutMapping(value = "/onl/onlinespec/modifyishavefirstorder")
-	Ro modifyIsHaveFirstOrderById(@RequestParam("id") Long id,
-			@RequestParam("isHaveFirstOrder") Boolean isHaveFirstOrder);
+    /**
+     * 根据上线规格id修改是否已有首单
+     * 
+     * @param id
+     * @param isHaveFirstOrder
+     * @return
+     */
+    @PutMapping(value = "/onl/onlinespec/modifyishavefirstorder")
+    Ro modifyIsHaveFirstOrderById(@RequestParam("id") Long id,
+            @RequestParam("isHaveFirstOrder") Boolean isHaveFirstOrder);
 }
